backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
pandas==2.2.3
openpyxl==3.1.5
requests==2.32.3
apscheduler==3.11.0
python-dateutil==2.9.0.post0
holidays==0.63
